requirements.txt
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-jwt==1.19.0
psycopg2==2.8.6
PyJWT==1.7.1
drf-yasg==1.20.0
Faker==8.2.1
